+++ framework/Source/NSCoderExtensions.m
@@ -237,11 +237,11 @@ void MyCGPathApplierFunc(void *info, const CGPathElement *element)
         [newKey release];
 
         for ( size_t i = 0; i < numberOfComponents; i++ ) {
-            newKey = [[NSString alloc] initWithFormat:@"%@.decode[%u].lower", key, (unsigned)i];
+            newKey = [[NSString alloc] initWithFormat:@"%@.decode[%zu].lower", key, i];
             [self encodeCGFloat:decodeArray[i * 2] forKey:newKey];
             [newKey release];
 
-            newKey = [[NSString alloc] initWithFormat:@"%@.decode[%u].upper", key, (unsigned)i];
+            newKey = [[NSString alloc] initWithFormat:@"%@.decode[%zu].upper", key, i];
             [self encodeCGFloat:decodeArray[i * 2 + 1] forKey:newKey];
             [newKey release];
         }
@@ -383,7 +383,7 @@ void MyCGPathApplierFunc(void *info, const CGPathElement *element)
 
     // decode elements
     for ( NSUInteger i = 0; i < count; i++ ) {
-        newKey = [[NSString alloc] initWithFormat:@"%@[%u].type", key, (unsigned)i];
+        newKey = [[NSString alloc] initWithFormat:@"%@[%lu].type", key, (unsigned long)i];
         CGPathElementType type = [self decodeIntegerForKey:newKey];
         [newKey release];
 
@@ -391,18 +391,18 @@ void MyCGPathApplierFunc(void *info, const CGPathElement *element)
 
         switch ( type ) {
             case kCGPathElementAddCurveToPoint: // 3 points
-                newKey = [[NSString alloc] initWithFormat:@"%@[%u].point3", key, (unsigned)i];
+                newKey = [[NSString alloc] initWithFormat:@"%@[%lu].point3", key, (unsigned long)i];
                 point3 = [self decodeCPTPointForKey:newKey];
                 [newKey release];
 
             case kCGPathElementAddQuadCurveToPoint: // 2 points
-                newKey = [[NSString alloc] initWithFormat:@"%@[%u].point2", key, (unsigned)i];
+                newKey = [[NSString alloc] initWithFormat:@"%@[%lu].point2", key, (unsigned long)i];
                 point2 = [self decodeCPTPointForKey:newKey];
                 [newKey release];
 
             case kCGPathElementMoveToPoint:    // 1 point
             case kCGPathElementAddLineToPoint: // 1 point
-                newKey = [[NSString alloc] initWithFormat:@"%@[%u].point1", key, (unsigned)i];
+                newKey = [[NSString alloc] initWithFormat:@"%@[%lu].point1", key, (unsigned long)i];
                 point1 = [self decodeCPTPointForKey:newKey];
                 [newKey release];
                 break;
@@ -493,11 +493,11 @@ void MyCGPathApplierFunc(void *info, const CGPathElement *element)
         decodeArray = malloc( numberOfComponents * 2 * sizeof(CGFloat) );
 
         for ( size_t i = 0; i < numberOfComponents; i++ ) {
-            newKey             = [[NSString alloc] initWithFormat:@"%@.decode[%u].lower", key, (unsigned)i];
+            newKey             = [[NSString alloc] initWithFormat:@"%@.decode[%zu].lower", key, i];
             decodeArray[i * 2] = [self decodeCGFloatForKey:newKey];
             [newKey release];
 
-            newKey                 = [[NSString alloc] initWithFormat:@"%@.decode[%u].upper", key, (unsigned)i];
+            newKey                 = [[NSString alloc] initWithFormat:@"%@.decode[%zu].upper", key, i];
             decodeArray[i * 2 + 1] = [self decodeCGFloatForKey:newKey];
             [newKey release];
         }
